+++ src/agent/CHANGELOG.md
@@ -8,6 +8,8 @@ CHANGELOG
  * [BC BREAK] Change AgentProcessor `keepToolMessages` to `excludeToolMessages` and default behaviour to preserve tool messages
  * Add `MetaDataAwareTrait` to `MockResponse`, the metadata will also be set on the returned `TextResult` when calling the `toResult` function
  * Add `HasSourcesTrait` to `Symfony\AI\Agent\Toolbox\Tool\Subagent`
+ * Add human-in-the-loop tool confirmation system
+ * Add `ToolCallRequested` event dispatched before tool execution
 
 0.3
 ---